requirements.txt
absl-py==2.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.55.3
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.67.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
Pillow==9.5.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
pyspng==0.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
